+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## Upcoming 0.5.0
+
+### Added
+
+- Added [`AsyncAccept::until`] method, that creates a new `AsyncAccept` that will stop accepting connections after another future finishes.
+
+### Changed
+- **Backwards incompatible**: `AsyncAccept::poll_accept` now returns, `Poll<Option<Result<...>>>` instead of `Poll<Result<...>>`. This allows the incoming stream of connections to stop, for example, if a graceful shutdown has been initiated. `impl`s provided by this crate have been updated, but custom implementations of `AsyncAccept`, or direct usage of the trait may break.
+
 ## 0.4.2 - 2022-04-09
 
 ### Added